refactor(groups): split groups model
