+++ src/app/main/sidebar/contact/groups/group_item.rs
@@ -8,11 +8,11 @@ use gtk::{Align, Box, Label, ListBox, ListBoxRow, Orientation, Picture};
 
 use tokio::task;
 
+use super::GroupsMsg;
 use crate::db::fs::{download_group_avatar_file, get_group_avatar_path};
 use crate::db::sql::Group;
-use super::ContactMsg;
 
-impl FactoryComponent<ListBox, ContactMsg> for Group {
+impl FactoryComponent<ListBox, GroupsMsg> for Group {
     type InitParams = Group;
     type Widgets = ();
     type Input = ();
